requirements.txt
colorama==0.3.7
contextlib2==0.5.1
defusedxml==0.4.1
Django==1.9.4
django-allauth==0.25.2
django-bootstrap-dynamic-formsets==0.4.5
django-bootstrap3==7.0.1
django-crispy-forms==1.6.0
django-debug-toolbar==1.4
django-extensions==1.6.1
django-filter==0.13.0
django-grappelli==2.8.1
django-guardian==1.4.4
django-mptt==0.8.3
django-registration==2.0.4
djangorestframework==3.3.3
httpie==0.9.3
Markdown==2.6.6
oauthlib==1.0.3
Pillow==3.2.0
psycopg2==2.6.1
Pygments==2.1.3
python3-openid==3.0.10
raven==5.12.0
requests==2.10.0
requests-oauthlib==0.6.1
six==1.10.0
sqlparse==0.1.19